# use Int64 for time