--- a/src/Swarm/Game/CESK.hs
+++ b/src/Swarm/Game/CESK.hs
@@ -83,6 +83,7 @@ module Swarm.Game.CESK (
 import Control.Lens ((^.))
 import Control.Lens.Combinators (pattern Empty)
 import Data.Aeson (FromJSON, ToJSON)
+import Data.Int (Int64)
 import Data.IntMap.Strict (IntMap)
 import Data.IntMap.Strict qualified as IM
 import GHC.Generics (Generic)
@@ -102,15 +103,15 @@ import Swarm.Util.WindowedCounter (Offsettable (..))
 
 -- | A newtype representing a count of ticks (typically since the
 --   start of a game).
-newtype TickNumber = TickNumber {getTickNumber :: Integer}
+newtype TickNumber = TickNumber {getTickNumber :: Int64}
   deriving (Eq, Ord, Show, Read, Generic, FromJSON, ToJSON)
 
 -- | Add an offset to a 'TickNumber'.
-addTicks :: Integer -> TickNumber -> TickNumber
-addTicks i (TickNumber n) = TickNumber $ n + i
+addTicks :: Int -> TickNumber -> TickNumber
+addTicks i (TickNumber n) = TickNumber $ n + fromIntegral i
 
 instance Offsettable TickNumber where
-  offsetBy = addTicks . fromIntegral
+  offsetBy = addTicks
 
 instance Pretty TickNumber where
   pretty (TickNumber i) = pretty i