--- a/src/Swarm/Game/Step.hs
+++ b/src/Swarm/Game/Step.hs
@@ -1037,7 +1037,7 @@ execConst c vs s k = do
       [VInt d] -> do
         time <- use $ temporal . ticks
         purgeFarAwayWatches
-        return $ Waiting (addTicks d time) (Out VUnit s k)
+        return $ Waiting (addTicks (fromIntegral d) time) (Out VUnit s k)
       _ -> badConst
     Selfdestruct -> do
       destroyIfNotBase $ \case False -> Just AttemptSelfDestructBase; _ -> Nothing
@@ -1461,7 +1461,7 @@ execConst c vs s k = do
       return $ Out (VDir (fromMaybe (DRelative DDown) $ mh >>= toDirection)) s k
     Time -> do
       TickNumber t <- use $ temporal . ticks
-      return $ Out (VInt t) s k
+      return $ Out (VInt $ fromIntegral t) s k
     Drill -> case vs of
       [VDir d] -> doDrill d
       _ -> badConst
@@ -1988,7 +1988,7 @@ execConst c vs s k = do
 
             -- Now wait the right amount of time for it to finish.
             time <- use $ temporal . ticks
-            return $ Waiting (addTicks (fromIntegral numItems + 1) time) (Out VUnit s k)
+            return $ Waiting (addTicks (numItems + 1) time) (Out VUnit s k)
       _ -> badConst
     -- run can take both types of text inputs
     -- with and without file extension as in
@@ -2231,7 +2231,7 @@ execConst c vs s k = do
         return $ Out v s k
       else do
         time <- use $ temporal . ticks
-        return . (if remTime <= 1 then id else Waiting (addTicks remTime time)) $
+        return . (if remTime <= 1 then id else Waiting (addTicks (fromIntegral remTime) time)) $
           Out v s (FImmediate c wf rf : k)
    where
     remTime = r ^. recipeTime